# backend/media-microservices/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Media Microservices API
  version: 1.0.0
servers:
  - url: /api
paths:
  /v1/video:
    post:
      operationId: uploadVideo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, contentMediaType: application/octet-stream}
      responses:
        '201':
          description: Created
          content:
            application/json: {schema: {$ref: '#/components/schemas/Video'}}
        '422': {$ref: '#/components/responses/Problem'}
    get:
      operationId: listVideos
      responses:
        '200':
          description: Newest first
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Video'}}
  /v1/video/{videoId}:
    get:
      operationId: getVideo
      parameters:
        - {$ref: '#/components/parameters/VideoId'}
      responses:
        '200':
          description: Video and renditions
          content:
            application/json: {schema: {$ref: '#/components/schemas/VideoDetails'}}
        '404': {$ref: '#/components/responses/Problem'}
  /v1/rendition:
    post:
      operationId: createRenditions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [videoId, resolutions]
              properties:
                videoId: {type: string, format: uuid}
                resolutions:
                  type: array
                  minItems: 1
                  uniqueItems: true
                  items: {$ref: '#/components/schemas/VideoResolution'}
      responses:
        '200': {description: Renditions persisted as PENDING}
        '400': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /v1/rendition/video/{videoId}:
    get:
      operationId: listRenditions
      parameters:
        - {$ref: '#/components/parameters/VideoId'}
      responses:
        '200':
          description: Empty for an unknown video
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Rendition'}}
  /v1/rendition/{renditionId}/download:
    get:
      operationId: downloadRendition
      parameters:
        - name: renditionId
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Finished rendition attachment
          headers:
            Content-Disposition: {schema: {type: string}}
          content:
            application/octet-stream: {schema: {type: string, format: binary}}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /v1/notification:
    get:
      operationId: listNotifications
      responses:
        '200':
          description: Newest first
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Notification'}}
  /v1/notification/stream:
    get:
      operationId: streamNotifications
      responses:
        '200':
          description: SSE events named notification
          content:
            text/event-stream:
              schema: {type: string}
components:
  parameters:
    VideoId:
      name: videoId
      in: path
      required: true
      schema: {type: string, format: uuid}
  responses:
    Problem:
      description: RFC 9457 Problem Detail
      content:
        application/problem+json: {schema: {$ref: '#/components/schemas/ProblemDetail'}}
  schemas:
    VideoResolution:
      type: string
      enum: [SD_360, SD_480, HD_720, FHD_1080]
    RenditionStatus:
      type: string
      enum: [PENDING, RUNNING, FINISHED, CANCELED, ERROR]
    VideoMetadata:
      type: object
      required: [videoCodec, containerFormat, width, height, durationSeconds, sizeBytes, bitRateBps]
      properties:
        videoCodec: {type: string, enum: [H264]}
        containerFormat: {type: string, enum: [MP4]}
        width: {type: integer}
        height: {type: integer}
        durationSeconds: {type: number}
        sizeBytes: {type: integer, format: int64}
        bitRateBps: {type: integer, format: int64}
    Video:
      type: object
      required: [id, name, metadata, createdAt]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        metadata: {$ref: '#/components/schemas/VideoMetadata'}
        createdAt: {type: string, format: date-time}
    VideoDetails:
      allOf:
        - {$ref: '#/components/schemas/Video'}
        - type: object
          required: [renditions]
          properties:
            renditions: {type: array, items: {$ref: '#/components/schemas/Rendition'}}
    Rendition:
      type: object
      required: [id, name, resolution, status, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        resolution: {$ref: '#/components/schemas/VideoResolution'}
        status: {$ref: '#/components/schemas/RenditionStatus'}
        error: {type: [string, 'null']}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    Notification:
      type: object
      required: [id, videoId, type, message, createdAt]
      properties:
        id: {type: string, format: uuid}
        videoId: {type: string, format: uuid}
        renditionId: {type: [string, 'null'], format: uuid}
        type: {type: string, enum: [TASK_COMPLETED, TASK_FAILED, ALL_COMPLETED]}
        message: {type: string}
        createdAt: {type: string, format: date-time}
    ProblemDetail:
      type: object
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string, format: uri}